Fix Windows build warnings treated as errors
- Fixed EntityID to int conversion warnings in event.h
- Fixed parameter shadowing class members in ai_system.h and equipment_system.h
- Fixed size_t to int conversions in fov.cpp and entity_factory.cpp
- Added explicit static_cast for all implicit conversions

All tests pass locally on macOS. These changes fix the CI build failures on Windows.

+++ include/ecs/event.h
@@ -137,7 +137,7 @@ inline BaseEvent DropEvent(EntityID dropper, EntityID item, int x, int y) {
 
 inline BaseEvent UseItemEvent(EntityID user, EntityID item, EntityID target = 0) {
     BaseEvent e(EventType::USE_ITEM, user, target ? target : user);
-    e.value1 = item;
+    e.value1 = static_cast<int>(item);
     return e;
 }
